\name{Chromatogram}
\alias{Chromatogram}

\docType{data}
\title{Example data of lysophosphatidic acids, LPA.}

\description{Data set containing times of detector hits after chromatographic separation of 
five different lysophosphatidic acids (Classes CLs = LPA 16:0, 18:0, 18:3, 20:0, and 20:4).}

\usage{data("Chromatogram")}

\details{
Size 1166 x 3 , stored in \code{Chromatogram$[Cls, Time, Lipids]}
}

\examples{
data(Chromatogram)
str(Chromatogram)
}
