\name{Mixture3}
\alias{Mixture3}

\docType{data}
\title{Example data of olfactory subtest performance.}

\description{Data set containing 1000 instances ditributed according to a Gaussian mixture with m = [-10, 0, 10], s = [1, 2, 3], w = [0.07, 0.05, 0.88].}

\usage{data("Mixture3")}

\details{
Size 1000 x 1
}

\examples{
data(Mixture3)
str(Mixture3)
}
